11. Prime factors 2

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -7,7 +7,7 @@
 
 int main(void)
 {
-	long long int i = 3, n = 612852475143;
+	unsigned long int i = 3, n = 612852475143;
 
 	for (; i < 12057; i += 2)
 	{
